requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
attrs==23.1.0
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.2.0
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
gunicorn==21.2.0
importlib-metadata==7.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tomli==2.0.1
typing_extensions==4.9.0
uuid==1.30
Werkzeug==3.0.1
zipp==3.17.0
